Complete all tasks from FINAL_ANALYSIS_SUMMARY.md - cleanup, fixes, enhancements
Major changes:
- Fixed 49-day timer overflow (uint32_t to uint64_t in heartbeat structures)
- Implemented last updated timer with color-based staleness indicators
- Enhanced debug page with current level display
- Removed duplicate build flags and cleaned up platformio.ini
- Added SSE real-time streaming for cell monitor
- Improved bar graph visualization (single-line flexbox, 120px height)
- Removed broken data source toggle from dashboard
- Added transmitter state indicator to hub page
- Cleaned legacy test mode code from receiver
- Added comprehensive documentation for all implementations

All projects build successfully without errors.

+++ ESP32common/espnow_transmitter/espnow_common.h
@@ -178,7 +178,7 @@ typedef struct __attribute__((packed)) {
 typedef struct __attribute__((packed)) {
     uint8_t  type;          // msg_heartbeat
     uint32_t seq;           // Monotonic sequence number (detects loss/resets)
-    uint32_t uptime_ms;     // Sender uptime in milliseconds
+    uint64_t uptime_ms;     // Sender uptime in milliseconds (64-bit to avoid overflow after 49 days)
     uint64_t unix_time;     // Unix timestamp in seconds (NTP-synchronized)
     uint8_t  time_source;   // 0=unsynced, 1=NTP, 2=manual, 3=GPS
     uint8_t  state;         // Connection state enum
@@ -191,7 +191,7 @@ typedef struct __attribute__((packed)) {
 typedef struct __attribute__((packed)) {
     uint8_t  type;          // msg_heartbeat_ack
     uint32_t ack_seq;       // Sequence number being acknowledged
-    uint32_t uptime_ms;     // Receiver uptime in milliseconds
+    uint64_t uptime_ms;     // Receiver uptime in milliseconds (64-bit to avoid overflow after 49 days)
     uint8_t  state;         // Receiver connection state
     uint16_t checksum;      // CRC16 checksum
 } heartbeat_ack_t;
